+++ app/export.go
@@ -71,9 +71,6 @@ func (app *TokenFactoryApp) prepForZeroHeightGenesis(ctx sdk.Context, jailAllowe
 		allowedAddrsMap[addr] = true
 	}
 
-	/* Just to be safe, assert the invariants on current state. */
-	app.CrisisKeeper.AssertInvariants(ctx)
-
 	/* Handle fee distribution state. */
 
 	// withdraw all validator commission